Move `get_column_specs` to its own module
