+++ databuilder/validate_dummy_data.py
@@ -1,21 +1,14 @@
 import csv
-import dataclasses
 import datetime
 import functools
 import gzip
 
 from databuilder import query_language as ql
-from databuilder.query_model import get_series_type
+from databuilder.column_specs import get_column_specs
 
 
 class ValidationError(Exception):
     pass
-
-
-@dataclasses.dataclass(frozen=True)
-class ColumnSpec:
-    type: type  # noqa: A003
-    nullable: bool = True
 
 
 def validate_file_types_match(dummy_filename, output_filename):
@@ -50,18 +43,6 @@ def get_file_type(filename):
         gzipped = False
     extension = suffixes[-1] if suffixes else ""
     return extension, gzipped
-
-
-def get_column_specs(variable_definitions):
-    # TODO: It may not be universally true that IDs are ints. Internally the EMIS IDs
-    # are SHA512 hashes stored as hex strings which we convert to ints. But we can't
-    # guarantee always to be able to pull this trick.
-    column_specs = {"patient_id": ColumnSpec(int, nullable=False)}
-    for name, series in variable_definitions.items():
-        if name == "population":
-            continue
-        column_specs[name] = ColumnSpec(get_series_type(series), nullable=True)
-    return column_specs
 
 
 def validate_csv_against_spec(csv_file, column_specs):
